Change instance number from int to long

+++ clients/SMR/rsrc/thrift/paxos.thrift
@@ -8,13 +8,13 @@ struct Value {
 
 struct Decision {
   1: i32 ring,
-  2: i32 instance,
+  2: i64 instance,
   3: Value value,
 }
 
 service PaxosProposerService {
 
-   i32 propose(1:Value value),
+   i64 propose(1:Value value),
    
    void nb_propose(1:Value value),
 }
@@ -25,7 +25,7 @@ service PaxosLearnerService {
 
    Decision nb_deliver(),
 
-   void safe(1:i32 ring,2:i32 instance),
+   void safe(1:i32 ring,2:i64 instance),
 }
